Spread out project structure
