+++ .env.example
@@ -4,5 +4,8 @@ POSTGRES_USER=postgres
 POSTGRES_PASSWORD=postgres
 
 # These must be loaded with python-dotenv.
-DB_URI="host=localhost dbname=postgres user=postgres password=postgres"
-TEST_DB_URI="host=localhost dbname=postgres user=postgres password=postgres port=5433"
+FLASK_ENV=development
+
+DB_CONN_STRING="host=localhost dbname=postgres user=postgres password=postgres"
+TEST_DB_CONN_STRING="host=localhost dbname=postgres user=postgres password=postgres port=5433"
+
