+++ .gitignore
@@ -2,4 +2,4 @@
 venv/
 .data/
 .vscode/
-**/__pycache__/
+__pycache__/
