+++ api/routes/__init__.py
@@ -1,0 +1,2 @@
+# flake8: noqa
+from . import patient, question, registration, station
