+++ docker-compose.yml
@@ -9,8 +9,6 @@ services:
       POSTGRES_USER: $POSTGRES_USER
       POSTGRES_PASSWORD: $POSTGRES_PASSWORD
     restart: unless-stopped
-    volumes:
-      - ./.data:/var/lib/postgresql/data
 
   test_db:
     image: postgres:latest
