+++ wsgi.py
@@ -1,0 +1,6 @@
+from api import create_app
+
+app = create_app()
+
+if __name__ == "__main__":
+    app.run(host="0.0.0.0", port=5000)
